+++ design-system/calendar/src/__tests__/unit/index.tsx
@@ -9,11 +9,6 @@ import Calendar, { type CalendarProps } from '../../index';
 import dateToString from '../../internal/utils/date-to-string';
 import { type TabIndex, type WeekDay } from '../../types';
 
-jest.mock('react-uid', () => ({
-	...jest.requireActual('react-uid'),
-	useUIDSeed: () => () => 'react-uid',
-}));
-
 const makeHandlerObject = ({
 	day,
 	month,